Add Current Education Level Field
I added a link to the jquery library and the code needed to access the
API with the data for the radio buttons and place then in <input> tags
that are appended to the label with the id of 'education'.

I also changed the css for label to be label[type="text"] so the radio
buttons revert to the natural style which looks better on the page.