Make Layout Responsive for Mobile Devices
I added in two media queries for 700px and 480px widths so everythings
looks good on mobile devices.

As part of that, I also added in support for older IE browsers with the
mediaqueries js.

I also added in more browser support for a couple CSS3 properties that
were used such as box-shadow and box-sizing.