removed an unnecessary div element from the education inputs
added an id to each label for use in validation.

added a call to the function validate for each label's onblur event, except for the education radio buttons, radio buttons are checked only when an atempt is made to submit the form.

added a call to the the function beforeSubmit.

when validation fails, the class "error" is added to the label, causing the css to display the title as a message to the user.

before submitting the the form, we run through the form one last time to verify each input is filed out and that no label has the class "error".